+++ Airbrake/gcalertview/GCAlertView.h
@@ -24,9 +24,6 @@
 
 #import <TargetConditionals.h>
 #if TARGET_OS_IPHONE
-#ifndef __IPHONE_4_0
-#error This project uses features only available in iOS 4.0 and later
-#endif
 #import <UIKit/UIKit.h>
 
 @interface GCAlertView : UIAlertView <UIAlertViewDelegate> {