+++ Airbrake/notifier/ABNotice.h
@@ -39,7 +39,7 @@ typedef struct ab_signal_info_t {
     void *user_data;
 	
 } ab_signal_info_t;
-ab_signal_info_t ab_signal_info;
+extern ab_signal_info_t ab_signal_info;
 
 // notice payload keys
 extern NSString * const ABNotifierOperatingSystemVersionKey;
@@ -56,6 +56,7 @@ extern NSString * const ABNotifierExceptionParametersKey;
 
 // notice file extension
 extern NSString * const ABNotifierNoticePathExtension;
+extern NSString * const ABNotifierExceptionPathExtension;
 
 // file flags
 extern const int ABNotifierNoticeVersion;
@@ -83,13 +84,12 @@ extern const int ABNotifierExceptionNoticeType;
     NSDictionary *__environmentInfo;
     NSArray *__callStack;
     NSNumber *__noticeVersion;
+    NSString *__userName;
 }
 
 // create an object representation of notice data
 - (id)initWithContentsOfFile:(NSString *)path;
 + (ABNotice *)noticeWithContentsOfFile:(NSString *)path;
-
-// get a string representation of the hoptoad xml payload
-- (NSString *)hoptoadXMLString;
-
+- (NSData *)JSONString;
+- (void)setPOSTUserName:(NSString *)theUserName;
 @end