+++ Xcode/Airbrake iOS/main.m
@@ -1,0 +1,16 @@
+//
+//  main.m
+//  Hoptoad iOS
+//
+//  Created by Caleb Davenport on 5/10/11.
+//  Copyright 2011 GUI Cocoa, LLC. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+#import "HTSAppDelegate_iOS.h"
+int main(int argc, char *argv[])
+{
+    @autoreleasepool {
+        return UIApplicationMain(argc, argv, nil, NSStringFromClass([HTSAppDelegate_iOS class]));
+    }
+}